Add files via upload
This is my project that let's you enter city names and temperatures of the city. It then bubble sorts all the citys and temps to show the lowest first, it also uses linear search to find specific temps. It also has functions that finds the Highest and the Lowest temps